Refactor task module
Split the previous `task.ts` into multiple files for better readability

Change-type: patch
